+++ package.json
@@ -1,68 +1,68 @@
 {
-	"name": "dms-ui",
-	"version": "1.1.3",
-	"private": true,
-	"scripts": {
-		"clearNextTemp": "rm -rf ./.next",
-		"dev": "npm run clearNextTemp && next dev",
-		"build": "npm run clearNextTemp && next build",
-		"patchTS": "ts-patch install -s",
-		"prepare": "npm run patchTS",
-		"start": "next start",
-		"test": "jest"
-	},
-	"dependencies": {
-		"@emotion/react": "^11.9.0",
-		"@emotion/styled": "^11.8.1",
-		"@overture-stack/arranger-components": "^3.0.0-beta.33",
-		"@overture-stack/dms-jbrowse-components": "^0.1.0-beta.3",
-		"@overture-stack/sqon-builder": "^0.0.0",
-		"axios": "^0.27.2",
-		"jsonwebtoken": "^8.5.1",
-		"jwt-decode": "^3.1.2",
-		"lodash": "^4.17.21",
-		"next": "^12.1.6",
-		"next-compose-plugins": "^2.2.1",
-		"next-global-css": "1.3.1",
-		"next-transpile-modules": "^9.0.0",
-		"query-string": "^7.1.1",
-		"react": "^17.0.2",
-		"react-dom": "^17.0.2",
-		"react-grid-system": "^7.1.1",
-		"react-rnd": "^10.4.1",
-		"react-tippy": "^1.4.0",
-		"url-join": "^5.0.0"
-	},
-	"devDependencies": {
-		"@babel/preset-typescript": "^7.21.0",
-		"@emotion/babel-plugin": "^11.9.2",
-		"@types/lodash": "^4.14.182",
-		"@types/node": "^17.0.35",
-		"@types/react": "^17.0.63",
-		"@types/react-dom": "^17.0.20",
-		"@types/url-join": "^4.0.1",
-		"extra-watch-webpack-plugin": "^1.0.3",
-		"fork-ts-checker-webpack-plugin": "^8.0.0",
-		"jest": "^27.5.1",
-		"prettier-plugin-organize-imports": "^3.2.3",
-		"ts-patch": "^2.1.0",
-		"typescript": "^4.9.5",
-		"typescript-transform-paths": "^3.4.6"
-	},
-	"prettier": {
-		"printWidth": 100,
-		"trailingComma": "all",
-		"semi": true,
-		"singleQuote": true,
-		"useTabs": false
-	},
-	"overrides": {
-		"node-fetch": "^2.6.1",
-		"graphiql": "1.5.17",
-		"react": "^17.0.2",
-		"react-dom": "^17.0.2"
-	},
-	"engines": {
-		"npm": ">=8.3.0"
-	}
+  "name": "dms-ui",
+  "version": "1.1.3",
+  "private": true,
+  "scripts": {
+    "clearNextTemp": "rm -rf ./.next",
+    "dev": "npm run clearNextTemp && next dev",
+    "build": "npm run clearNextTemp && next build",
+    "patchTS": "ts-patch install -s",
+    "prepare": "npm run patchTS",
+    "start": "next start",
+    "test": "jest"
+  },
+  "dependencies": {
+    "@emotion/react": "^11.9.0",
+    "@emotion/styled": "^11.8.1",
+    "@overture-stack/arranger-components": "^3.0.0-beta.33",
+    "@overture-stack/dms-jbrowse-components": "^0.1.0-beta.3",
+    "@overture-stack/sqon-builder": "^0.0.0",
+    "axios": "^0.27.2",
+    "jsonwebtoken": "^8.5.1",
+    "jwt-decode": "^3.1.2",
+    "lodash": "^4.17.21",
+    "next": "^12.1.6",
+    "next-compose-plugins": "^2.2.1",
+    "next-global-css": "1.3.1",
+    "next-transpile-modules": "^9.0.0",
+    "query-string": "^7.1.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-grid-system": "^7.1.1",
+    "react-rnd": "^10.4.1",
+    "react-tippy": "^1.4.0",
+    "url-join": "^5.0.0"
+  },
+  "devDependencies": {
+    "@babel/preset-typescript": "^7.21.0",
+    "@emotion/babel-plugin": "^11.9.2",
+    "@types/lodash": "^4.14.182",
+    "@types/node": "^17.0.35",
+    "@types/react": "^17.0.63",
+    "@types/react-dom": "^17.0.20",
+    "@types/url-join": "^4.0.1",
+    "extra-watch-webpack-plugin": "^1.0.3",
+    "fork-ts-checker-webpack-plugin": "^8.0.0",
+    "jest": "^27.5.1",
+    "prettier-plugin-organize-imports": "^3.2.3",
+    "ts-patch": "^2.1.0",
+    "typescript": "^4.9.5",
+    "typescript-transform-paths": "^3.4.6"
+  },
+  "prettier": {
+    "printWidth": 100,
+    "trailingComma": "all",
+    "semi": true,
+    "singleQuote": true,
+    "useTabs": true
+  },
+  "overrides": {
+    "node-fetch": "^2.6.1",
+    "graphiql": "1.5.17",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2"
+  },
+  "engines": {
+    "npm": ">=8.3.0"
+  }
 }
